[>] [me_Display_Mono128x64] Imported from [Warf]..
Well, this code is new and it works.

Problem is that I don't have projects for it to test. So no point
of releasing it yet.

Putting fresh code in unmaintained decaying repository is not good.
Probably, I'll split this repo to something.

Generally [Scrapyard] repository contains three types of code:

  * Experimental
  * Historic
    * For personal memories
    * For potential release